% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_SB_WB.R
\name{error.test}
\alias{error.test}
\title{Nonparametric Two-Sample Test for Error Distributions}
\usage{
error.test(x1, y1, x2, y2, methods = c("SB", "WB"))
}
\arguments{
\item{x1}{Numeric vector representing the predictor values for the first sample.}

\item{y1}{Numeric vector representing the response values for the first sample.}

\item{x2}{Numeric vector representing the predictor values for the second sample.}

\item{y2}{Numeric vector representing the response values for the second sample.}

\item{methods}{Character vector specifying the method(s) to use, either 'SB', 'WB',
or both. Default is c('SB', 'WB').}
}
\value{
A list containing information about the test including sample sizes (n1, n2)
and the p-value.
}
\description{
This function performs a two-sample test for comparing error distributions
in nonparametric regression. It supports two methods: Weighted Bootstrap (WB)
and Smooth Bootstrap (SB).
}
\examples{
n1 = 39
n2 = 49
x1 = runif(n1)
y1 = runif(n1)
x2 = runif(n2)
y2 = runif(n2)
error.test(x1, y1, x2, y2, methods = 'SB')
error.test(x1, y1, x2, y2, methods = 'WB')
}
\references{
Kernel Density Estimation and Its Applications, G. S. Watson, 1983.
}
